Implements downloader tool

+++ tools/cirdl/cirdl.nim
@@ -4,6 +4,7 @@ import pkg/chronos
 import pkg/ethers
 import pkg/questionable
 import pkg/questionable/results
+import pkg/zip/zipfiles
 import ../../codex/contracts/marketplace
 
 ## TODO: chronicles is still "Log message not delivered: [Chronicles] A writer was not configured for a dynamic log output device"
@@ -28,13 +29,27 @@ proc getCircuitHash(rpcEndpoint: string, marketplaceAddress: string): Future[?!s
   return success config.proofs.zkeyHash
 
 proc formatUrl(hash: string): string =
-  return "A"
+  "https://circuit.codex.storage/proving-key/" & hash
 
-proc downloadZipfile(url: string, filepath: string) =
-  info "a"
+proc downloadZipfile(url: string, filepath: string): ?!void =
+  try:
+    # Nim's default webclient does not support SSL on all platforms.
+    # Not without shipping additional binaries and cert-files... :(
+    # So we're using curl for now.
+    var rc = execShellCmd("curl -o " & filepath & " " & url)
+    if not rc == 0:
+      return failure("Download of '" & url & "' failed with return code: " & $rc)
+  except Exception as exc:
+    return failure(exc.msg)
+  success()
 
-proc unzip(zipfile:string, targetPath: string) = 
-  info "a"
+proc unzip(zipfile:string, targetPath: string): ?!void = 
+  var z: ZipArchive
+  if not z.open(zipfile):
+    return failure("Unable to open zip file: " & zipfile)
+  z.extractAll(targetPath)
+  z.close()
+  success()
 
 proc main() {.async.} =
   info "Codex Circuit Downloader, Aww yeah!"
@@ -45,19 +60,30 @@ proc main() {.async.} =
 
   let
     circuitPath = args[0]
-    rpcEndpoint = "http://kubernetes.docker.internal:30001" #args[1]
-    marketplaceAddress = "0x111F5aAA5DFF76510b220d152426fe878B4a87AE" #args[2]
+    rpcEndpoint = args[1]
+    marketplaceAddress = args[2]
     zipfile = circuitPath / "circuit.zip"
+
+  debug "Starting", circuitPath, rpcEndpoint, marketplaceAddress
+
+  discard existsOrCreateDir(circuitPath)
 
   without circuitHash =? (await getCircuitHash(rpcEndpoint, marketplaceAddress)), err:
     error "Failed to get circuit hash", msg = err.msg
     return
-  info "got circuitHash", circuitHash
+  debug "Got circuithash", circuitHash
 
   let url = formatUrl(circuitHash)
+  if dlErr =? downloadZipfile(url, zipfile).errorOption:
+    error "Failed to download circuit file", msg = dlErr.msg
+    return
+  debug "Download completed"
 
-  downloadZipfile(url, zipfile)
-  unzip(zipfile, circuitPath)
+  if err =? unzip(zipfile, circuitPath).errorOption:
+    error "Failed to unzip file", msg = err.msg
+    return
+  debug "Unzip completed"
+
   removeFile(zipfile)
 
 when isMainModule:
